fastapi_app/requirements.txt
fastapi
uvicorn
psycopg2-binary
celery
redis
jinja2
python-multipart
requests